+++ CHANGELOG.md
@@ -9,6 +9,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 ### Added
 
 - Add ability to import GIAS groups from csv data
+- transfer project now collect if they are a result of 2RI and show this
+  information on the about the project tab
 
 ## [Release-46][release-46]